@@ -1,9 +1,8 @@
 ext.versions = [
-        compileSdkVersion: 28,
+        compileSdkVersion: 29,
         minSdkVersion: 16,
         targetSdkVersion: 28,
 
-        librarySupport: '28.0.0',
         libraryFirebase: [
                 analytics: '17.2.0',
                 perf: '19.0.1'
