@@ -2,68 +2,47 @@ package com.wt.pinger.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.lifecycle.ViewModelProviders;
 
-import com.crashlytics.android.answers.Answers;
-import com.crashlytics.android.answers.ContentViewEvent;
 import com.google.firebase.perf.metrics.AddTrace;
 import com.hivedi.era.ERA;
-import com.pnikosis.materialishprogress.ProgressWheel;
-import com.wt.pinger.App;
 import com.wt.pinger.R;
+import com.wt.pinger.proto.StartActivityObserver;
+import com.wt.pinger.proto.StartViewModel;
 
-public class StartActivity extends AppCompatActivity implements App.OnAppReady {
+public class StartActivity extends AppCompatActivity {
 
-    private ProgressWheel progress;
-
-    private final Runnable startRun = new Runnable() {
-        @Override
-        public void run() {
-            if (mApp.isAppReady()) {
-                progress.stopSpinning();
-                progress.setVisibility(View.INVISIBLE);
-                startActivity(new Intent(StartActivity.this, MainActivity.class));
-                finish();
-                overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out);
-            } else {
-                mApp.setOnAppReady(StartActivity.this);
-            }
-        }
-    };
-    private App mApp;
+    //private ProgressWheel progress;
 
     @Override
     @AddTrace(name = "StartActivity_onCreate")
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         ERA.log("StartActivity.onCreate:begin");
-        mApp = (App) getApplication();
         setContentView(R.layout.activity_start);
-        progress = findViewById(R.id.progress_wheel);
+        //progress = findViewById(R.id.progress_wheel);
+
+        StartViewModel model = ViewModelProviders.of(this).get(StartViewModel.class);
+        model.getData().observe(this, new StartActivityObserver(this));
+        model.getData().setValue(
+            new Intent(StartActivity.this, MainActivity.class)
+        );
+
         ERA.log("StartActivity.onCreate:end");
     }
 
     @Override
     protected void onPause() {
         ERA.log("StartActivity.onPause");
-        progress.removeCallbacks(startRun);
         super.onPause();
     }
 
     @Override
     protected void onResume() {
         super.onResume();
-        Answers.getInstance().logContentView(
-                new ContentViewEvent()
-                        .putContentId("start-activity")
-                        .putContentName("Start Activity")
-                        .putContentType("activity")
-        );
         ERA.log("StartActivity.onResume");
-        progress.spin();
-        progress.post(startRun);
     }
 
     @Override
@@ -72,8 +51,4 @@ public class StartActivity extends AppCompatActivity implements App.OnAppReady {
         // nop
     }
 
-    @Override
-    public void onAppReady() {
-        progress.post(startRun);
-    }
 }
